requirements.txt
aiogram
aiohttp
aiosqlite
python-dotenv
cryptography
paramiko
GitPython